Implemented caching, fixed coordinates
Caching for the TileMap is now implemented (it remembers the most recent
appropriate 64x64 grid, and updates the cache in a sane way).

Coordinates are now appropriately implemented (upper-right in pixel
terms is positive Y movement in world terms, and lower-right in pixel
terms is positive X movement in world terms), and movement works
properly.