Moved keyboard scrolling into the Demo project
Realistically, it should never have been part of the engine.

Mouse controls are still part of it, though; moused-over squares are
still automagically selected, primarily because writing the code for it
was quite unpleasant and I wouldn't want an end-user to have to do that.

I've updated the readme to reflect this.