Added in tile overrides
There is now a THIRD layer of tiles in the TileMap, called overrides.
They are implemented as another layer of SavedCells, so to get a
specific tile, the map asks overrides, then savedcells, then cache, then
generates it new.

The purpose is for "in-world" UI elements to be quickly generated and
dropped (such as selecting a bunch of tiles, then visually replacing
them with some placeholder thing, a la zoning in Sim City).