Added FPS counters
One only need to create an FPSComponent object (in the initialize
phase), then load it with a font (in the LoadContent phase) and it will
take care of itself.

Note it will only appear if you put it on the top of your components, so
be aware of that.  Also, it can be toggled on and off by modfying
FPSComponent.Visible