[Moderate] Added dirty bits for pathfinding
Specifically, there is now a field in TileMapManager which keeps track
of the last time the passability system was altered (and a method for
saying "it was altered right now," dependent on implementation to use).

If the PathHunter detects that the TileMapManager has been changed in
this way since it was started, PathHunter will start over.