Improved MapCache, to allow for teleportation
The way the Cache was working, if you teleported the Camera to somewhere
weird (say, to an important event which was far away) the Cache would
react in a really inefficient way (basically walking over there instead
of just *being there*).  It now detects this situation, and if the
situation demands it, just rebuilds the entire cache from scratch
instead of moving smoothly.