[Minor] Fiddled previous patch
I was, um, holding it wrong.

Rectangle DOES work as (they) intended, and IS designed for ints.  It's
just defined by slightly different inequalities than I thought.  (x,y)
is in the rectangle iff Left <= x < Right and Top <= y < Bottom

(not the mixed strict and non-strict inequalities; this makes for
natural for-loops but somewhat non-intuitive definitions for
non-programmers)