Camera tweaks, made TileMapManager generic
Camera now has a center on point method.  It requires that you remember
to size your TileMap appropriately (which in turn sets the Camera size
appropriately!) but is otherwise not especially exciting.

Also, TileMapManager is now generic; the basic InGameObjects are now a
type.  This allows one to require all InGameObjects to be some extension
class, guaranteeing extra functionality (e.g. new tags) without a lot of
casting issues.