[Performance] MapCache is now for visual cells
Again, this is a significant performance improvement.  The real cells
typically need to be called from all over the place, so do not fit the
current implementation of the MapCache well.  However, the visual cells
are really, really rectangular.  It works great there, so now we can
really use the benefits of the cache.