Added cell borders and limiting [BREAKS OLD CODE]
This is not a huge release, but it does break some old code.  I'm not
worried about that since no one is using this code (yet?) but, you know.

Two things:

1) TileMaps can now be limited.  As of now, this allows settings min/min
x and y coordinates, outside of which nothing will be DRAWN.  This
doesn't change anything else, though.  Not a big change.  As time goes
on, this will either expand or go away, I predict.

2) Cells can now have borders, which are independent and
four-directional.  This requires 5 new boolean flags and 4 new integers,
and the demo code shows a basic nontrivial usage.  Note this is a major
performance hit (drawing up to 4 new images per tile!).  While it may
not be a huge deal, depending on usage, if you're in optimize mode,
start here.